% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mead1988Data.R
\name{Abrasion}
\alias{Abrasion}
\title{Abrasion data}
\format{
A data frame with observations on the following variables.
\describe{
}
}
\source{
Data found in table 13.28 on page 380.
}
\usage{
Abrasion
}
\description{
Data from a 2^5 experiment.
}
\references{
Mead, R. (1988) \emph{The Design of Experiments: Statistical Principles for Practical Applications}, Cambridge University Press.
}
\author{
contributed by A. Jonathan R. Godfrey
}
