% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CochranCox1957.R
\name{Beans}
\alias{Beans}
\title{Factorial experiment for Bean Yield}
\source{
Table 6.3 on page 189.
}
\usage{
Beans
}
\description{
A 2^4 experiment in two blocks; blocks are confounded with the four-way interaction, but there are two replicates within each block.
}
\references{
Cochran, W.G. and Cox, G.M.  (1957) \emph{Experimental Designs}, Second edition, John Wiley & Sons, New York.
}
\author{
contributed by Jonathan Godfrey
}
