% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeanVosseData.R
\name{BeefTender}
\alias{BeefTender}
\title{Beef Tenderness Data}
\source{
adapted from files found at \url{http://deanvossdraguljic.ietsandbox.net/DeanVossDraguljic/R-data.html}
}
\usage{
BeefTender
}
\description{
Data for the Beef Tenderness Experiment originally reported by Cochran and Cox (1957). The response variable is
}
\references{
Cochran, W.G. and Cox, G.M.  (1957) \emph{Experimental Designs}, Second edition, John Wiley & Sons, New York.

Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.

Dean, Voss, and Draguljić (2017) \emph{Design and Analysis of Experiments}, Second edition, , Springer-Verlag New York.
}
\author{
Contributed by A. Jonathan R. Godfrey
}
