% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mead1988Data.R
\name{CabbageFlies}
\alias{CabbageFlies}
\title{CabbageFlies data}
\format{
A data frame with observations on the following variables.
\describe{
}
}
\source{
Mead, R. (1988) "The Design of Experiments: Statistical Principles for Practical Applications", Cambridge University Press. Example 7.10; data found in  Table 7.3 on page 145.
}
\usage{
CabbageFlies
}
\description{
Data from a ... experiment
}
\author{
contributed by A. Jonathan R. Godfrey
}
