% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualEngData.R
\name{CakeMix}
\alias{CakeMix}
\title{Cake Mix Data}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with observations on the following variables.
}\if{html}{\out{</div>}}
}
\source{
to fix
}
\usage{
CakeMix
}
\description{
Cake Mix Data
}
\author{
contributed by A. Jonathan R. Godfrey
}
