% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CochranCox1957.R
\name{CC1957}
\alias{CC1957}
\title{Data from Cochran and Cox (1957)}
\description{
Datasets from Cochran and Cox (1957) are found in a variety of places. Not all sources state which tables they have used from the original source so some digging was required.
}
\details{
The data has been cleaned if it is mentioned here.
}
\references{
Cochran, W.G. and Cox, G.M.  (1957) \emph{Experimental Designs}, Second edition, John Wiley & Sons, New York.
}
\seealso{
BeefTender, DairyCow, Field
}
\author{
contributed by Jonathan Godfrey
}
