% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CochranCox1957.R
\name{CC57_Beans}
\alias{CC57_Beans}
\title{Factorial experiment for Bean Yield}
\format{
A data.frame with 32 observations. Variables are \code{Block}, \code{Yield}, and the four treatment factors.
Factors are \code{D} = Dung, \code{N}=Nitrochalk, \code{P}=Superphosphate, and \code{K}=Muriate of potash.
All were Absent (-1) or present (1).
}
\source{
Table 6.3 on page 189.
}
\usage{
CC57_Beans
}
\description{
A 2^4 experiment in two blocks; blocks are confounded with the four-way interaction, but there are two replicates within each block.
}
\references{
Cochran, W.G. and Cox, G.M.  (1957) \emph{Experimental Designs}, Second edition, John Wiley & Sons, New York.
}
\author{
contributed by A. Jonathan R. Godfrey
}
