% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CochranCox1957.R
\name{CC57_Cotton}
\alias{CC57_Cotton}
\title{CC57_Cotton}
\format{
A data.frame with 32 observations. Variables are \code{Block}, \code{Yield}, and the four treatment factors.
}
\source{
Table 11.3 on page 456.
}
\usage{
CC57_Cotton
}
\description{
CC57_Cotton
}
\references{
Cochran, W.G. and Cox, G.M.  (1957) \emph{Experimental Designs}, Second edition, John Wiley & Sons, New York.
}
\author{
contributed by A. Jonathan R. Godfrey
}
