% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualEngData.R
\name{Drilling}
\alias{Drilling}
\title{Drilling Data}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with observations on the following variables.
}\if{html}{\out{</div>}}
}
\source{
https://doi.org/10.1080/08982112.2011.535696
}
\usage{
Drilling
}
\description{
Drilling Data
}
\author{
contributed by A. Jonathan R. Godfrey
}
