% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualEngData.R
\name{Ethanol}
\alias{Ethanol}
\title{Ethanol Data}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with observations on the following variables.
}\if{html}{\out{</div>}}
}
\source{
\url{https://doi.org/10.1080/08982112.2013.817013}
}
\usage{
Ethanol
}
\description{
Ethanol Data
}
\author{
contributed by A. Jonathan R. Godfrey
}
