% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mead1988Data.R
\name{GLCompetition}
\alias{GLCompetition}
\title{Grass-Legume Competition data}
\format{
A data frame with observations on the following variables.
\describe{
}
}
\source{
Mead, R. (1988) \emph{The Design of Experiments: Statistical Principles for Practical Applications}, Cambridge University Press. Data found in table 14.15 on page 418.
}
\usage{
GLCompetition
}
\description{
Data from a split plot experiment. Three factors tested at the whole plot level are grass type, legume type, and cutting regime. The proportion of grass to legume is the subplot factor.
}
\author{
contributed by A. Jonathan R. Godfrey
}
