% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualEngData.R
\name{GrandmasCake}
\alias{GrandmasCake}
\title{Grandma's Cake Mix Data}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with eight observations on the following variables.
}\if{html}{\out{</div>}}

Factors for Egg (A), Sugar (B), and Flour (C);
the run order;
a blocking variable, Block;
the Rating of each recipe.
}
\source{
\url{https://doi.org/10.1080/08982119708919097}
}
\usage{
GrandmasCake
}
\description{
A 2^3 experiment run in two blocks using the abc interaction to define blocks.
}
\author{
contributed by A. Jonathan R. Godfrey
}
