% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualEngData.R
\name{InaTile}
\alias{InaTile}
\title{Taguchi's Ina Tile Data}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with observations on the following variables.
}\if{html}{\out{</div>}}
}
\source{
\url{https://doi.org/10.1081/QEN-120020782}
}
\usage{
InaTile
}
\description{
Taguchi's Ina Tile Data
}
\author{
contributed by A. Jonathan R. Godfrey
}
