% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kuehl2000Data.R
\name{Kuehl2000}
\alias{Kuehl2000}
\title{Data from Kuehl 2000}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{All data is stored as a data frame. N.B. Original filenames included minus signs which have been converted to underscores. In order ot remove confusion with data from other texts, all data names start with Kuehl_, e.g. EX2-1.csv is now Kuehl_EX2_1  
}\if{html}{\out{</div>}}
}
\source{
Kuehl, R.O. (2000) \emph{Design of Experiments: Statistical Principles of Research Design and Analysis}, Second edition, Duxbury Press.
}
\description{
Data for many examples and exercises from Kuehl (2000) was obtained in ASCII format and has been prepared for use as R data. More rigorous preparation will be required for many of the datasets.
}
\author{
contributed by A. Jonathan R. Godfrey
}
