% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualEngData.R
\name{Lenses}
\alias{Lenses}
\title{Lenses Data}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with observations on the following variables.
}\if{html}{\out{</div>}}
}
\source{
to fix
}
\usage{
Lenses
}
\description{
Lenses Data
}
\author{
contributed by A. Jonathan R. Godfrey
}
