% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeanVosseData.R
\name{NorthStar}
\alias{NorthStar}
\title{North Star Steel Iowa Data on manganese}
\source{
Dean and Vosse
}
\usage{
NorthStar
}
\description{
North Star Steel Iowa Data on manganese
}
\author{
Contributed by A. Jonathan R. Godfrey
}
