% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mead1988Data.R
\name{PainRelief}
\alias{PainRelief}
\title{PainRelief data}
\format{
A data frame with observations on the following variables.
\describe{
}
}
\source{
Mead, R. (1988) "The Design of Experiments: Statistical Principles for Practical Applications", Cambridge University Press. Example 8.2; data found in Table 8.8 on page 200.
}
\usage{
PainRelief
}
\description{
Data from a ... experiment
}
\author{
contributed by A. Jonathan R. Godfrey
}
