% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mead1988Data.R
\name{Peppers}
\alias{Peppers}
\title{Peppers data}
\format{
A data frame with 24 observations on the following 6 variables.
\describe{
\item{\code{Heating}}{a factor with levels \code{Extra} \code{Standard}}
\item{\code{Lighting}}{a factor with levels \code{Extra} \code{Standard}}
\item{\code{CO2}}{a factor with levels \code{Added} \code{c} \code{Control}}
\item{\code{Year}}{a factor with levels \code{1} \code{2}}
\item{\code{Block}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
\item{\code{Excess}}{a numeric vector}
}
}
\source{
Mead, R. (1988) "The Design of Experiments: Statistical Principles for Practical Applications", Cambridge University Press. Example 7.11; data found in Table 7.7 on page 148.
}
\usage{
Peppers
}
\description{
Data from a ... experiment
}
\author{
contributed by A. Jonathan R. Godfrey
}
