% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{Std8Runs}
\alias{Std8Runs}
\alias{Std16Runs}
\alias{Std32Runs}
\title{Unreplicated 2^k Experiments in Standard (Yates) Order}
\format{
a dataframe with a column for each factor and 2^k rows, where k=3,4,5
}
\source{
Take your pick.
}
\description{
Each dataset has factors A, B, C, ... coded as -1 and 1 for the low and high levels of each two-level factor. Examples exist for three, four, and five factors.
}
\author{
created by A. Jonathan R. Godfrey, to whom no credit for these datasets  is justified.
}
