% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeanVosseData.R
\name{SteelBarCutting}
\alias{SteelBarCutting}
\title{Steel Bar Cutting Data}
\source{
adapted from files found at \url{http://deanvossdraguljic.ietsandbox.net/DeanVossDraguljic/R-data.html}
}
\usage{
SteelBarCutting
}
\description{
Table 7.26 of Dean and Vosse (1999).
}
\references{
Dean, A. and Voss, D. (1999). \emph{Design and Analysis of Experiments}. New York, Springer.
Dean, Voss, and Draguljić (2017) \emph{Design and Analysis of Experiments}, Second edition, , Springer-Verlag New York.
}
\author{
Contributed by A. Jonathan R. Godfrey
}
