% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{Taguchi8}
\alias{Taguchi8}
\alias{Taguchi16#'}
\alias{@format}
\alias{a}
\alias{dataframe}
\alias{with}
\alias{column}
\alias{for}
\alias{each}
\alias{factor}
\alias{and}
\alias{2^k}
\alias{rows,}
\alias{all}
\alias{-1}
\alias{columns}
\alias{the}
\alias{full}
\alias{set}
\alias{of}
\alias{main}
\alias{interaction}
\alias{effects}
\alias{,}
\alias{where}
\alias{k=3}
\alias{or}
\alias{4}
\title{Unreplicated 2^k Experiments in Taguchi form}
\source{
Take your pick.
}
\description{
Each structure has columns c1, c2, ...coded as -1 and 1 for the low and high levels of each two-level factor.
}
\author{
created by A. Jonathan R. Godfrey, to whom no credit for these datasets  is justified.
}
