% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mead1988Data.R
\name{Turnip}
\alias{Turnip}
\title{Turnip data}
\format{
A data frame with observations on the following variables.
\describe{
}
}
\source{
Mead, R. (1988) "The Design of Experiments: Statistical Principles for Practical Applications", Cambridge University Press. Example 12.3; data found in  Table 12.9 on page 323.
}
\usage{
Turnip
}
\description{
Data from a ... experiment
}
\author{
contributed by A. Jonathan R. Godfrey
}
