% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mead1988Data.R
\name{Vaccine}
\alias{Vaccine}
\title{Vaccine data}
\format{
A data frame with observations on the following variables.
\describe{
}
}
\source{
Mead, R. (1988) "The Design of Experiments: Statistical Principles for Practical Applications", Cambridge University Press. Example 11.1; data found in table 11.7 on page 289.
}
\usage{
Vaccine
}
\description{
Data from a ... experiment
}
\author{
contributed by A. Jonathan R. Godfrey
}
