% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualEngData.R
\name{VoidCount}
\alias{VoidCount}
\title{VoidCount Data}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with observations on the following variables.
}\if{html}{\out{</div>}}
}
\source{
https://doi.org/10.1080/08982110600846531
}
\usage{
VoidCount
}
\description{
VoidCount Data
}
\author{
contributed by A. Jonathan R. Godfrey
}
