% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualEngData.R
\name{Washing}
\alias{Washing}
\title{Washing Data}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with observations on the following variables.
}\if{html}{\out{</div>}}
}
\source{
\url{https://doi.org/10.1081/QEN-200029007}
}
\usage{
Washing
}
\description{
Washing Data
}
\author{
contributed by A. Jonathan R. Godfrey
}
