% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bhh2005Data.R
\name{Wool}
\alias{Wool}
\title{Wool Data from BHH2}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with 27 observations on the following five variables.
}\if{html}{\out{</div>}}

\describe{
\item{\code{X}}{a numeric vector}
\item{\code{Length}}{a numeric vector}
\item{\code{Amplitude}}{a numeric vector}
\item{\code{Load}}{a numeric vector}
\item{\code{Cycles}}{a numeric vector}
}
}
\source{
A digital copy was obtained as No. 402 of Hand et al. (1994).
}
\usage{
Wool
}
\description{
The number of cycles before failure of a worsted yarn under cycles of repeated loading to three factors arranged in a 3^3 factorial experiment.
}
\references{
Box, G.E.P., Hunter, J.S., and Hunter, W.G. (2005) \emph{Statistics for Experimenters: Design Innovation, and Discovery}, 2nd edition, Wiley, New Jersey.
G.E.P. and D.R. Cox (1964) \dQuote{An Analysis of Transformations (with discussion)}, \emph{Journal of the Royal Statistical Society, Series B} \bold{26}, pp.211--246.
}
\author{
contributed by A. Jonathan R. Godfrey
}
