% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mead1988Data.R
\name{Detergent}
\alias{Detergent}
\title{Detergent data}
\format{
A data frame with observations on the following variables.
\describe{
}
}
\source{
Mead, R. (1988) "The Design of Experiments: Statistical Principles for Practical Applications", Cambridge University Press. Example 12.4; data found in  Table 12.13 on page 329.
}
\usage{
Detergent
}
\description{
Data from a ... experiment
}
\author{
contributed by A. Jonathan R. Godfrey
}
