% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualEngData.R
\name{Drilling}
\alias{Drilling}
\title{Drilling Data}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with observations on the following variables.
}\if{html}{\out{</div>}}
}
\source{
https://doi.org/10.1080/08982112.2011.535696
}
\usage{
Drilling
}
\description{
2^4 factorial experiment that studies the effects of drill load (factor A), flow rate (factor B), rotational speed (factor C), and type of mud (factor D)
on the advance rate for an oil drill.
}
\references{
Montgomery DAE problem 15.2 (Example 6.3), QualEng, Daniel
}
\author{
contributed by A. Jonathan R. Godfrey
}
