% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mead1988Data.R
\name{PainRelief}
\alias{PainRelief}
\title{Pain Relief data}
\format{
A data frame with 86 observations on the following four variables.
\describe{
\item{\code{Patient}}{a factor with levels \code{1} to \code{43}}
\item{\code{Period}}{a factor with levels \code{1} \code{2}}
\item{\code{Drug}}{a factor with levels \code{T1} \code{T2} \code{T3}}
\item{\code{Hours}}{the response variable}
}
}
\source{
Mead, R. (1988) "The Design of Experiments: Statistical Principles for Practical Applications", Cambridge University Press. Example 8.2; data found in Table 8.8 on page 200.
}
\usage{
PainRelief
}
\description{
Data from a ... experiment
}
\author{
contributed by A. Jonathan R. Godfrey
}
