% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{Taguchi8}
\alias{Taguchi8}
\alias{Taguchi16}
\title{Unreplicated 2^k Experiments in Taguchi form}
\format{
a dataframe with a column for each factor and 2^k rows, all 2^k -1 columns for the full set of main and interaction effects , where k=3 or 4
}
\source{
Take your pick.
}
\description{
Each structure has columns c1, c2, ...coded as -1 and 1 for the low and high levels of each two-level factor.
}
\author{
created by A. Jonathan R. Godfrey, to whom no credit for these datasets  is justified.
}
